fixed a bug which could lead to interger overflows for big optimizables

+++ bundles/org.palladiosimulator.simexp.dsl.ea.optimizer/src/main/java/org/palladiosimulator/simexp/dsl/ea/optimizer/smodel/PowerUtil.java
@@ -50,10 +50,10 @@ public class PowerUtil {
     }
 
     public int getPowerRangeInt(RangeBounds rangeBounds) {
-        int startValue = expressionCalculator.calculateInteger(rangeBounds.getStartValue());
-        int endValue = expressionCalculator.calculateInteger(rangeBounds.getEndValue());
-        int stepSize = expressionCalculator.calculateInteger(rangeBounds.getStepSize());
-        int power = Math.round(((endValue * 1000000) - (startValue * 1000000)) / (stepSize * 1000000));
+        double startValue = expressionCalculator.calculateInteger(rangeBounds.getStartValue());
+        double endValue = expressionCalculator.calculateInteger(rangeBounds.getEndValue());
+        double stepSize = expressionCalculator.calculateInteger(rangeBounds.getStepSize());
+        int power = (int) Math.round(((endValue * 1000000) - (startValue * 1000000)) / (stepSize * 1000000));
         return power;
     }
